### 2020-07-10 #2471 @whedon
[PRE REVIEW]: NetworkChange: Analyzing Network Changes in R
**Submitting author:** @ysohn (<a href="http://orcid.org/0000-0002-9496-1907">Yunkyu Sohn</a>)
**Repository:** <a href="https://github.com/jongheepark/NetworkChange" target ="_blank">https://github.com/jongheepark/NetworkChange</a>
**Version:** v0.6
