+++ R/NetworkStatic.R
@@ -77,8 +77,8 @@
 #'
 #' @seealso \code{\link{NetworkChange}}
 #'
-#' @references   Jong Hee Park and Yunkyun Sohn. 2019. "Detecting Structural Change
-#' in Longitudinal Network Data." \emph{Bayesian Analysis}. Forthcoming.
+#' @references   Jong Hee Park and Yunkyun Sohn. 2020. "Detecting Structural Change
+#' in Longitudinal Network Data." \emph{Bayesian Analysis}. Vol.15, No.1, pp.133-157.
 #'
 #' Peter D. Hoff 2011. "Hierarchical Multilinear Models for Multiway Data."
 #' \emph{Computational Statistics \& Data Analysis}. 55: 530-543.
