+++ R/drawpostanalysis.R
@@ -10,7 +10,7 @@
 #' @param n.cluster number of cluster. Default is 3. 
 #'
 #' @references   Jong Hee Park and Yunkyun Sohn. 2020. "Detecting Structural Change
-#' in Longitudinal Network Data." \emph{Bayesian Analysis}. Forthcoming.
+#' in Longitudinal Network Data." \emph{Bayesian Analysis}. Vol.15, No.1, pp.133-157.
 #' 
 #' @return A plot object
 #'
