### 2020-07-13 #2471 @kakiac
👋  @ysohn thanks for your submission to JOSS! 🥳

I think I would like us to work on @kyleniemeyer 's comments above: https://github.com/openjournals/joss-reviews/issues/2471#issuecomment-656894068 before I pass this on to reviewers to review, as it will substantially improve the submission before they review it. Will you be able to address them?

See for example this paper for an example of what a JOSS paper looks like: https://joss.theoj.org/papers/10.21105/joss.02265

Many thanks and looking forward to working with you on this submission,
