### 2020-07-14 #2471 @kakiac
👋  @akbaritabar @marcwilliams1 @martinmodrak @ejolly would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
