### 2020-07-14 #2471 @ejolly
@kakiac Unfortunately I don't think I'll be that helpful for this review, since it's be quite a while since I've worked natively with R (predominantly python these days). Happy to attempt it though if no other reviewers are available. @kakiac Is it also possible to edit my "Other Languages" on the joss review list to remove R and Matlab? Would be good so I'm not a time bottleneck on future submissions that require reviewers for those language.
