### 2020-07-16 #2471 @martinmodrak
@kakiac Hi, the package looks interesting and I'd be glad to review it. Unfortunately, I am currently leaving for vacation until July 26th and will probably have a bunch of stuff on my plate after the return. So realistically, I might not have an initial review ready sooner than Aug 1st. If that's OK, I can review it. (I am also OK if you look for other reviewers who can react faster and fall back on me only if that fails).
