### 2020-10-01 #2708 @whedon
[REVIEW]: NetworkChange: Analyzing Network Changes in R
**Submitting author:** @ysohn (<a href="http://orcid.org/0000-0002-9496-1907">Yunkyu Sohn</a>)
**Repository:** https://github.com/jongheepark/NetworkChange
**Branch with paper.md** (empty if default branch):
**Version:** v0.6
