### 2020-10-01 #2708 @kakiac
Editorial Comments carried over from [Pre-review] https://github.com/openjournals/joss-reviews/issues/2471 (@kakiac to check):

-[ ] Your software repository does not appear to include an open source license
-[ ] Your paper is missing an explicit Statement of Need section, which is now required for articles as described in our documentation
-[ ] Please remove these lines from your paper's YAML header: aas-doi: 10.3847/xxxxx and aas-journal: Journal of Open Source Software.
-[ ] Also, overall your paper is quite short. Although JOSS papers are generally only one or two pages on average (250-1000 words), your paper just barely passes that mark. I would recommend adding a bit more content, perhaps an example or more explanation/details of how the package works.
-[ ] It looks like a few of your references may be missing DOIs
