### 2020-10-01 #2471 @kakiac
@ejolly re this https://github.com/openjournals/joss-reviews/issues/2471#issuecomment-658327413 I have requested a change in the spreadsheet, so hopefully you will not be getting more requests on R!
