### 2020-10-01 #2708 @kakiac
👋🏼 @ysohn @akbaritabar, @marcjwilliams1, @martinmodrak this is the review thread for the paper. All of our communications will happen here from now on.

All reviewers should have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 6 weeks. Please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
