### 2020-10-25 #2708 @arfon
@akbaritabar, @marcjwilliams1, @martinmodrak - how are you getting on with your reviews here? Please note we try to _complete_ the review process in 6 weeks which means that it's helpful for you all to provide initial feedback to the authors earlier than this.
