### 2020-10-26 #2708 @marcjwilliams1
I don't seem to be able to check off items on the checklist, if I click through to https://github.com/openjournals/joss-reviews/invitations as suggested I get the following message
`Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.`
