### 2020-10-29 #2708 @arfon
@marcjwilliams1 - can you try clicking the invite link again now? I believe the invites expire after a week.
