### 2020-10-30 #2708 @akbaritabar
## Letter to authors
Thank you for your trust and allowing me to take part in this interesting experience as I am studying peer review as a subject of research and the way JOSS is handling it was really interesting to me.

I must start by saying that I am a social scientist who uses R, Python and network analysis tools and techniques on daily basis. That means I have evaluated the submission with the above background and most of my points below are not technical at all. Reading the package paper, vignette and the authors' 2020 paper in Bayesian Analysis journal, I was ready to give up on the review and state a lack of _technical_ confidence. Then I decided to act as an _advanced beginner_ user of the package and try to provide feedback on my experience that _might_ be useful for authors and the editor (I certainly hope so). I will state some main points that I faced and below some minor suggestions. Further below I have copied some of the points from the review checklist and added comments on them that I needed to describe.

A bit more description about the empirical network data used would be nice (e.g., I found this in the 2020 paper by the authors in Bayesian Analysis journal which made it much clearer for me (quoting this from that paper's text) "The structure of military alliance networks reflects the distribution of military power among coalitions of states, which changes over time in response to exogenous shocks to the international system or endogenous network dynamics. However, there has been no study that investigates changes in coalition structures of military alliance networks over time using a principled statistical approach. A main reason is the lack of statistical methods that model unknown structural changes in longitudinal network data.").

Largest possible size of the network and longitudinal waves that can be analyzed and visualized by the package is not mentioned in paper or vignette. For the current example in vignette, there was the "BreakDiagnostic" step that on my laptop (core i 7, 16GB RAM) took a few minutes to be done, I was wondering if authors can give estimates of possible time it can take and what factors can affect that in network attributes and analysis terms for users to have an idea.

An example of how objects from popular R libraries for network analysis (e.g., igraph and statnet suite) or external formats (e.g., Pajek plain text format) can be converted and used with NetworkChange will be really helpful. At first it feels like array of temporal snapshots is easy to build, but it might not be for all potential users. I took the time to test it on a fake co-authorship network and it was not as straightforward as I hoped it would be (please see above in first paragraph that I am an advanced beginner and it might only be my problem). I used a bipartite edgelist saved in CSV including source/target columns for authors and publication nodes and a publication year column (see below for the example). I constructed a bipartite graph with igraph, then induced yearly subgraphs (to have temporal dimension to explore with NetworkChange) and then projected them to one-mode (authors). I then used array from R to get close to the MajourAlly shape, but I failed and couldn't understand how I need to provide the dimensions to then use NetworkChange on the sample. Providing a bit of description on this would allow the package to be used by a wider range of audience.

This is example of bipartite ties I tried to convert and used and failed in the latest step:

```R

# example data (edges table) in csv format to copy/paste

source,target,PUBYEAR
author_1,paper_2,2007
author_2,paper_2,2007
author_3,paper_2,2007
author_4,paper_2,2007
author_5,paper_2,2007
author_1,paper_3,2011
author_2,paper_3,2011
author_3,paper_3,2011
author_4,paper_3,2011
author_5,paper_3,2011
author_1,paper_4,2012
author_2,paper_4,2012
author_3,paper_4,2012
author_4,paper_4,2012
author_5,paper_4,2012
author_1,paper_1,2013
author_2,paper_1,2013
author_3,paper_1,2013
author_4,paper_1,2013
author_5,paper_1,2013

# example data (vertices table) in csv format to copy/paste

vertices,type
author_1,TRUE
author_2,TRUE
author_3,TRUE
author_4,TRUE
author_5,TRUE
paper_1,FALSE
paper_2,FALSE
paper_3,FALSE
paper_4,FALSE

# Preparing to use with NetworkChange
require(tidyverse)
require(igraph)

edges_table <- read_csv('./edges.csv') %>%
  arrange(PUBYEAR)

vertices_table <- read_csv('./vertices.csv')

# build a bipartite paper-author graph
g <- graph_from_data_frame(d = edges_table, vertices = vertices_table, directed = F)

yearly_g <- list()
for (year in unique(edges_table$PUBYEAR)) {
  # delete.vertices = FALSE because we need same dim for all matrices
  g_year <- subgraph.edges(g, E(g)[PUBYEAR == year], delete.vertices = FALSE)
  g_aut <- bipartite.projection(g_year, which = 'true')
  g_adj <- as.matrix(get.adjacency(g_aut, edges = F, names = T))
  yearly_g[[paste0('y_', as.character(year))]] <- g_adj
}

# Here I am not sure how to give the right dimensions and fail in building the right array that NetworkChange will use
arr_yearly_g <- array(yearly_g)

### the error I get is like this:

# use networkchange
G <- 100
set.seed(1990)
test.run <- NetworkStatic(arr_yearly_g, R=2, mcmc=G, burnin=G, verbose=0,
                          v0=10, v1=4*2)

# Error in array(Y, dim = c(dim(Y)[1], dim(Y)[2], 1)) :
#   negative length vectors are not allowed

```

Below I present some minor points as suggestions to be revised in text. I am sorry that I don't have much technical suggestions to offer and I hope other reviewers and editor would provide enough comments for those aspects.

### Minor points
#### In JOSS paper
- "The complete guide for using core functions of the package is presented at https://github.com/jongheepark/NetworkChange as its vignette with an empir- ical data set analysis example." -> This is the link to repository not the vignette (https://cran.r-project.org/web/packages/NetworkChange/vignettes/NetworkChange.html)
- In both paper and the vignette (see below), figure 1 ("Summary of selected features and functions of the package.") is used but in text there no description of it. It is a nice figure summarizing the main features and function names, it would be nicer to have a few sentences describe it (IMO).
- "library(NetworkChange)" is missing from the first code chunk of the paper (it is present in the vignette).
- "We follow the COW dataset’s coding of “major powers” (the United Kingdom, Germany, Austria-Hungary, France, Italy, Russia, the United States, Japan, and China) in the analysis. We aggregated every 2 year network from the original annual binary networks to increase the density of each layer." -> in the script accompanying this part you are dropping CHN and USA since they don't have ties to other nodes, but in text it is not mentioned that they are dropped.
- - "Users can choose the number of clusters in each regime by ‘n.cluster}" -> "` `" instead of "}"

#### In Vignette
- "Users can change plot settings by changing options in `ggnet`." -> maybe, but changing options which _are passed to_ `ggnet`?
- "In this section, We estimate the chagepoints of" -> changepoints
- "We also highlight the latent cluster structure using the k-means clustering method over the estimated latent node positions of each regime. drawPostAnalysis() provides an option to include k-means clustering results of the latent node positions for each regime." -> I felt the two sentences were similar.
- "In this section, we analyze changes in the international military alliance network among major powers. The data set is originally from (???)" -> reference in parenthesis is missing from the vignette.
- "Our goal in this section is to detect structural changes in the longitudinal alliance network among major powers using HNC. We follow the COW dataset’s coding of “major powers” (the United Kingdom, Germany, Austria-Hungary, France, Italy, Russia, the United States, Japan, and China) in the analysis. We aggregated every 2 year network from the original annual binary networks to increase the density of each layer." -> in the script accompanying this part you are dropping CHN and USA since they don't have ties to other nodes, but in text it is not mentioned that they are dropped.
- "Users can choose the number of clusters in each regime by ‘n.cluster}" -> "` `" instead of "}"
- Figure captions present in the paper are missing from the vignette, but I am not sure if the vignette would support them or not? If it does, it would be nice to have them here as well.
- "International military alliances, 1648-2008. (2009). Washington, DC: CQ Press. doi:10.4135/ 9781604265781" -> Reference to empirical data used is missing from vignette

#### In package documentation PDF from CRAN
- "orthgonalization" -> "orthogonalization"

#### Points from Review checklist that I needed to elaborate
- General checks
    - [x] Contribution and authorship: Has the submitting author (@ysohn) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
        - The submitting author seems to have joined recently (https://github.com/jongheepark/NetworkChange/graphs/contributors) but the first author have worked on this repository from late 2017. The current order of authorship seems to respect that IMO.
    - [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines
        - There are two peer reviewed publications (2017 and 2020) based on the methodology as far as I can see and the lead author is the maintainer of CRAN task view on Bayesian analysis, I feel confident the contribution would be good. But I might not be technically qualified to judge that.
- Functionality
        - Installation from CRAN and latest version from Github went fine on Mac OS.
        - But, once installed from github, the vignette of package is not found in R with ` vignette(package = "NetworkChange") no vignettes found `
- Documentation
    - [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
        - Vignette and paper gives empirical example and scripts to see how functions work. But I am not sure if I can find automated tests.
        - On the package repository or its Wiki there is no guide on this, but Github issue tracker and pull requests should be the intended ones, I suppose.
- Software paper
    - [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
        - The current description can be further improved. For example, for me as someone who knows about network analysis, block-modeling, community detection and similar techniques, I cannot say the whole methodology and descriptions are clear but that could be my lack of enough background knowledge.
    - [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
        - I have given minor suggestions for paper and the vignette below.
