### 2020-11-05 jongheepark/NetworkChange#3 @marcjwilliams1
Automated tests
related to openjournals/joss-reviews#2708

One of the checklist items is to have automated tests. Using testthat and github actions are easy ways to implement this.
