+++ vignettes/NetworkChange.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
-%% https://bibdesk.sourceforge.io/
+%% http://bibdesk.sourceforge.net/
 
-%% Created for Yunkyu Sohn at 2020-07-16 03:35:13 +0900 
+%% Created for 박종희 at 2020-11-05 23:29:56 +0900 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -9,15 +9,12 @@
 
 
 @book{Gibler2009,
-	Address = {Washington, DC},
-	Date-Added = {2020-07-16 03:35:12 +0900},
-	Date-Modified = {2020-07-16 03:35:12 +0900},
-	Doi = {10.4135/9781604265781},
+	Author = {Gibler, Douglas},
+	Date-Added = {2020-11-05 23:27:24 +0900},
+	Date-Modified = {2020-11-05 23:27:24 +0900},
 	Publisher = {CQ Press},
-	Title = {International Military Alliances, 1648-2008},
-	Year = {2009},
-	Bdsk-Url-1 = {http://sk.sagepub.com/cqpress/international-military-alliances-1648-2008},
-	Bdsk-Url-2 = {https://doi.org/10.4135/9781604265781}}
+	Title = {{International Military Alliances, 1648-2008}},
+	Year = {2008}}
 
 @article{Park2020,
 	Author = {Jong Hee Park and Yunkyu Sohn},
