### 2020-11-05 jongheepark/NetworkChange#4 @marcjwilliams1
Community guidelines
Related to openjournals/joss-reviews#2708

I couldn't find anything that fulfilled this:
```
Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
```

I think a short note in the README would suffice.
