### 2020-11-05 #2708 @marcjwilliams1
Overall everything looks fine, I get the same outputs when I run the vignette but would be nice to see some automated tests jongheepark/NetworkChange#3 . A note on community guidelines should also be added somewhere jongheepark/NetworkChange#4 .
