### 2020-11-05 #2708 @kakiac
👋🏼  @akbaritabar, @marcjwilliams1 many thanks for your reviews and comprehensive comments.

@ysohn can you please start addressing the reviewers' comments? I will be creating issues in your repo from @akbaritabar's comments to help us keep track of the comments to be addressed 😄

@martinmodrak I see that you have not checked any of your checklist items - can you edit it or are you having similar problems like @marcjwilliams1 ?
