### 2020-11-08 jongheepark/NetworkChange#5 @martinmodrak
No example to demonstrate fitting with covariates
Based on the description in the Bayesian analysis paper, fitting models with covariates should be supported. It is however unclear how one should do that as the `NetworkChange` function has parameters only for response and priors, not for predictors. The vignette or other documentation should show how to do this, or - if this is not currently possible (which is OK), then I don't understand what the prior for β  (parameters `b0` and `B0`) is for.
