### 2020-11-08 jongheepark/NetworkChange#6 @martinmodrak
Impossible to verify convergence, examples do not converge
In any MCMC run, it is important to verify that the chain actually converged. To my knowledge, the best practice is to run multiple chains and then compute one of the family of the R-hat diagnostics that compares within- and between- chain variance and also the effective sample size [1]. One could also show traceplots for visual inspection of mixing, but that is usually quite a weak test.

In fact the examples provided do not converge based on this criteria. Code to reproduce my tests is at https://gist.github.com/martinmodrak/b13077fe298f5e7567fe1975a2e2fcdd (it is a bit of a hack, obviously, in any practical setting, one would run the chains in parallel and maybe one could get away with just 4 chains.). In the example with `MakeBlockNetworkChange`, there are at least two separate problems:

1) Some of the `U` parameters have multimodal posteriors, possibly due to some label-switching or other symmetry (it seems that the sign of the parameters is flipped, see traceplot below). I can imagine this being resolvable by post processing, but I don't understand the underlying math enough to be sure. Or maybe users don't care about the actual `U` values? (if so, it should be documented that `U` values are not be taken too seriously)

![image](https://user-images.githubusercontent.com/9483603/98483405-43f2cb00-2208-11eb-80e6-2c598139029f.png)

2) The chains require more burnin samples to get to the stationary distribution - for the `s2` parameters (which are the state transition probabilities, if I get it right), the Rhat is about 1.08 but it should be < 1.01 and the effective sample size is about 73 (from 8 chains, 100 samples each = 800 samples total) which is barely enough to even estimate central tendencies (and this is from 8 times the number of samples used in the example). A traceplot below shows some mild issues, e.g. the "black dashed" chain howering somewhat above the others for the first 50 iterations or so and only then mixing with the rest.

![image](https://user-images.githubusercontent.com/9483603/98483592-0f800e80-220a-11eb-9c79-cb3dcd474d69.png)

Those problems don't seem to affect the qualitative inferences made _much_, although the individual chains make somewhat different plots, (showing only two chains for brevity)

![image](https://user-images.githubusercontent.com/9483603/98483762-4dc9fd80-220b-11eb-8702-bbce440f2b4e.png)
![image](https://user-images.githubusercontent.com/9483603/98483772-50c4ee00-220b-11eb-96a3-f5815d65084f.png)

![image](https://user-images.githubusercontent.com/9483603/98483750-43a7ff00-220b-11eb-9ded-da59e9481e1f.png)
![image](https://user-images.githubusercontent.com/9483603/98483751-46a2ef80-220b-11eb-9c2d-f0c9d2201d47.png)

For the ally network example, the multimodality in `U` persists, but we also get multimodality in the `s2` parameters, e.g. the traceplot below:

![image](https://user-images.githubusercontent.com/9483603/98483673-9c2acc80-220a-11eb-88c4-7c05d8f270e9.png)

This results in fundamentally different inferences about the transitions (showing only two chains once again):

![image](https://user-images.githubusercontent.com/9483603/98483836-a1d4e200-220b-11eb-9e8b-83553e52c968.png)
![image](https://user-images.githubusercontent.com/9483603/98483837-a6999600-220b-11eb-92dd-acdff61c26ea.png)

or for the overall clusters/closeness:

![image](https://user-images.githubusercontent.com/9483603/98483885-efe9e580-220b-11eb-8edf-66ace8f32baf.png)
![image](https://user-images.githubusercontent.com/9483603/98483901-098b2d00-220c-11eb-833a-aebaeed836ea.png)

# What to do about this?

Running multiple chains, computing diagnostics for all parameters of interest and issuing a warning if the diagnostics fail is IMHO necessary to be done by default to allow safe usage of the package by non-experts. I am not sure to what extent one would want `U` to behave neatly - maybe some sort of procrustes transform as a postprocessing step would mitigate the issues found in the `MakeBlockNetworkChange` example.

I am not sure if the ally dataset actually has multimodal posterior or if the chains would converge to a single mode given enough iterations, but this definitely bears investigating.

[1] Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.
