### 2020-11-08 jongheepark/NetworkChange#3 @martinmodrak
Just to add my thoughts on this: automated tests for MCMC algorithms are hard to do. In any case, I think ability to generate simulated data that follows exactly the model (i.e. draw `U`, `V`, `s` from their priors and then simulate the data from there) is necessary. Or maybe if this creates to uninteresting random graphs then maybe force some structure on `U` instead of drawing from the prior. But still actually simulating the HMM and having known "true" values for all parameters for which we make inference. We can then check the coverage of posterior credible intervals over those true values (e.g. how many "true" values fall within 50% and 95% credible intervals). I think for a basic test, this could be done "by hand" in a vignette/script. But in a best world, one would do a full Simulation-based calibration (https://arxiv.org/abs/1804.06788) which is AFAIK the strongest test for a Bayesian inference algorithm we currently have.

Hope I am making sense here, feel free to ask for details.
