### 2020-11-08 #2708 @martinmodrak
So I did a first round of review. I think I can complement @akbaritabar in that I've never done network analysis, but I have background in Bayesian inference and even some experience with fitting HMMs. I however have to be clear that my exposure to Bayesian methods is primarily via the Stan language and the associated community (e.g. Gelman, Vehtari, Betancourt, Simpson) which tends to have opinionated worldview on how Bayesian statistics should be done which may not necessarily match the consensus of the whole field. So feel free to challenge specific propositions/suggestions I make.

The problem attacked by the package is hard. My first impression upon reading the associated Bayesian analysis paper was "I wonder if this kind of model can be fit reliably with any available method.". To some extent my suspicion has been vindicated - I've noticed some problems that threaten the validity of the computation (https://github.com/jongheepark/NetworkChange/issues/6), so this is currently my biggest concern. In the case I am not wrong about the issues (which I might be), resolving them completely might turn out to be a significant research project on its own and I don't want to force the authors to do that. I would generally be happy if the package provides diagnostics and warnings to the user in case something looks fishy.

The paper itself (and the package) relies heavily on the Bayesian analysis paper for any deeper explanations. I am mostly OK with this, although I can see how adding at least rough description of the actual model to the JOSS paper would be beneficial. I will be happy if @kakiac as the editor weighs in if this is appropriate.
