### 2020-11-08 #2708 @martinmodrak
I should also add that installation was easy and following the examples was also easy, so definitely good job there.
