### 2020-11-09 jongheepark/NetworkChange#6 @martinmodrak
I probably also should add that I believe (though I am not sure and haven't tested this heavily) that the problem is - at least primarily - not the temporal part of the model, but rather I would guess the MTRM model would have multimodal posteriors for some data even when applied to a single time point (or multiple observations of a single time point).
