### 2020-11-09 jongheepark/NetworkChange#7 @kakiac
[JOSS #2708] Expand paper
(copying comments from https://github.com/openjournals/joss-reviews/issues/2708 )

Overall your paper is quite short. Although JOSS papers are generally only one or two pages on average (250-1000 words), your paper just barely passes that mark. I would recommend adding a bit more content, perhaps an example or more explanation/details of how the package works. Consider adding elements from this page: https://cran.r-project.org/web/packages/NetworkChange/vignettes/NetworkChange.html
