### 2020-11-09 jongheepark/NetworkChange#7 @kakiac
You might also consider adding something along the lines of from @akbaritabar's comment (https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111)

> A bit more description about the empirical network data used would be nice (e.g., I found this in the 2020 paper by the authors in Bayesian Analysis journal which made it much clearer for me (quoting this from that paper's text) "The structure of military alliance networks reflects the distribution of military power among coalitions of states, which changes over time in response to exogenous shocks to the international system or endogenous network dynamics. However, there has been no study that investigates changes in coalition structures of military alliance networks over time using a principled statistical approach. A main reason is the lack of statistical methods that model unknown structural changes in longitudinal network data.").
