### 2020-11-09 jongheepark/NetworkChange#8 @kakiac
[JOSS #2708] Largest possible size of the network and longitudinal waves 
Comment from @akbaritabar https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111
> Largest possible size of the network and longitudinal waves that can be analyzed and visualized by the package is not mentioned in paper or vignette. For the current example in vignette, there was the "BreakDiagnostic" step that on my laptop (core i 7, 16GB RAM) took a few minutes to be done, I was wondering if authors can give estimates of possible time it can take and what factors can affect that in network attributes and analysis terms for users to have an idea.
