### 2020-11-09 jongheepark/NetworkChange#10 @kakiac
[JOSS #2708] Fix link to vignette
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> The text mentions: "The complete guide for using core functions of the package is presented at https://github.com/jongheepark/NetworkChange as its vignette with an empir- ical data set analysis example."

> change requested: This is the link to repository not the vignette (https://cran.r-project.org/web/packages/NetworkChange/vignettes/NetworkChange.html)
