### 2020-11-09 jongheepark/NetworkChange#11 @kakiac
[JOSS #2708] Discuss figure 1 description in main text
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> In both paper and the vignette (see below), figure 1 ("Summary of selected features and functions of the package.") is used but in text there no description of it. It is a nice figure summarizing the main features and function names, it would be nicer to have a few sentences describe it (IMO).
