### 2020-11-09 jongheepark/NetworkChange#12 @kakiac
[JOSS #2708] Add library(NetworkChange) in code
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> "library(NetworkChange)" is missing from the first code chunk of the paper (it is present in the vignette).
