### 2020-11-09 jongheepark/NetworkChange#13 @kakiac
[JOSS #2708] Clarify assumptions around which major powers are included
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> "We follow the COW dataset’s coding of “major powers” (the United Kingdom, Germany, Austria-Hungary, France, Italy, Russia, the United States, Japan, and China) in the analysis. We aggregated every 2 year network from the original annual binary networks to increase the density of each layer." -> in the script accompanying this part you are dropping CHN and USA since they don't have ties to other nodes, but in text it is not mentioned that they are dropped.
