### 2020-11-09 jongheepark/NetworkChange#14 @kakiac
[JOSS #2708] Minor typos in code
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> "Users can choose the number of clusters in each regime by ‘n.cluster}" -> " " instead of "}"

i.e. remove the curly bracket after 'n.cluster'
