### 2020-11-09 #2708 @kakiac
👋  @ysohn, hope you are well, I have tried to organise the reviewers' comments up to now in separate issues in your repository to help us all track progress :). If you hover over them, you can see their status.

Hope it helps!

Many thanks to all reviewers for the comments up to now, @ysohn let us know if you need any help or further clarification on any of our comments.

Comments from **JOSS editors**:

- https://github.com/jongheepark/NetworkChange/pull/1
- https://github.com/jongheepark/NetworkChange/issues/7

Comments from **@marcjwilliams1**:
- https://github.com/jongheepark/NetworkChange/issues/3
- https://github.com/jongheepark/NetworkChange/issues/4

Comments from **@martinmodrak**:
- https://github.com/jongheepark/NetworkChange/issues/5
- https://github.com/jongheepark/NetworkChange/issues/6

Comments from **@akbaritabar**:
- https://github.com/jongheepark/NetworkChange/issues/8
- https://github.com/jongheepark/NetworkChange/issues/9
- https://github.com/jongheepark/NetworkChange/issues/10
- https://github.com/jongheepark/NetworkChange/issues/11
- https://github.com/jongheepark/NetworkChange/issues/12
- https://github.com/jongheepark/NetworkChange/issues/13
- https://github.com/jongheepark/NetworkChange/issues/14
