### 2020-11-09 jongheepark/NetworkChange#15 @kakiac
[JOSS #2708] Language improvements/typos to the vignette
From @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

Here are some suggestions for improving the vignette - these are not a pre-requisite for accepting the paper, but you might wish to consider them to improve the quality of the overall submission. I have added them as check items here in case that helps 😄

> ** In Vignette**

> - [x] "Users can change plot settings by changing options in ggnet." -> maybe, by changing options which are passed to ggnet?
> - [x] "In this section, We estimate the chagepoints of" -> changepoints
> - [x] "We also highlight the latent cluster structure using the k-means clustering method over the estimated latent node positions of each regime. drawPostAnalysis() provides an option to include k-means clustering results of the latent node positions for each regime." -> I felt the two sentences were similar.
> - [x] "In this section, we analyze changes in the international military alliance network among major powers. The data set is originally from (???)" -> reference in parenthesis is missing from the vignette.
> - [x] "Our goal in this section is to detect structural changes in the longitudinal alliance network among major powers using HNC. We follow the COW dataset’s coding of “major powers” (the United Kingdom, Germany, Austria-Hungary, France, Italy, Russia, the United States, Japan, and China) in the analysis. We aggregated every 2 year network from the original annual binary networks to increase the density of each layer." -> in the script accompanying this part you are dropping CHN and USA since they don't have ties to other nodes, but in text it is not mentioned that they are dropped.
> - [x] "Users can choose the number of clusters in each regime by ‘n.cluster}" -> " " instead of "}"
Figure captions present in the paper are missing from the vignette, but I am not sure if the vignette would support them or not? If it does, it would be nice to have them here as well.
> - [x] "International military alliances, 1648-2008. (2009). Washington, DC: CQ Press. doi:10.4135/ 9781604265781" -> Reference to empirical data used is missing from vignette

> **In package documentation PDF from CRAN**
> - [x] "orthgonalization" -> "orthogonalization"
