### 2020-11-09 jongheepark/NetworkChange#7 @kakiac
You might also consider improving/adding a few sentences to help a non-specialist audience understand what this package does and how it fits with other packages both in the readme.md of your repository and your paper. See also @akbaritabar 's comment here https://github.com/openjournals/joss-reviews/issues/2708#issuecomment-719592111:

> The current description can be further improved. For example, for me as someone who knows about network analysis, block-modeling, community detection and similar techniques, I cannot say the whole methodology and descriptions are clear but that could be my lack of enough background knowledge.
