### 2020-11-09 jongheepark/NetworkChange#4 @kakiac
Also noted by @akbaritabar here: https://github.com/jongheepark/NetworkChange/issues/16
