### 2020-12-01 #2708 @kakiac
@yson no problem at all. I will set an automatic reminder to all of us for mid January - in the meantime hope you all take some time to relax and "recharge" your batteries :)
