### 2021-01-31 #2708 @ysohn
@kakiac Dear editor, Happy new year. We are working to resolve the multimodality problem raised by one of the reviewers. It takes longer than expected as we have to experiment several candidates for identification of the MCMC samples. It'd be great if we could have more time to revise the package.
