### 2021-01-31 #2708 @martinmodrak
@ysohn: if that's in response to my review,  I reiterate that I would find it sufficient for publication in JOSS if the package provided diagnostics and clearly documented how a user can verify that the results can be trusted - or which parts of the results should be taken seriously (e.g. if the `U` parameters are not interpretable). My hunch is that for general datasets, the identification issues are irresolvable, so maybe good diagnostics to check if the model works for a given dataset is the best we can hope for. Obviously, if you can resolve or mitigate the issues,  its good for everyone,  but I didn't intend to send you on such a large project.
