### 2021-02-02 #2708 @ysohn
@martinmodrak  Thank you so much for your reply. I understand your point now. Without artificial constraints, factorized estimates in general have degrees of freedom coming from rotation and scaling. In our model, the sum of outer products of the estimates are meant to produce matrices that are eventually used in the model whereas the factorized estimates themselves are instrumental for this recovery. Thus, if we do not impose artificial constraints for the factorized estimates, the key for the diagnosis is to show the convergence up to the resulting matrices. We will work on this. Thank you for your patience.
