### 2021-11-30 jongheepark/NetworkChange#15 @ysohn
We have updated the vignette. Thanks!
