### 2021-11-30 jongheepark/NetworkChange#9 @ysohn
The following is included in the paper.

> The input network data for **NetworkChange** is a full three-dimensional array with dimensions $N$ by $N$ by $T$, where $N$ is the number of nodes and $T$ is the number of observed time points, and each slice represents an adjacency matrix for a specific time point. Most other network packages include a function for converting an edge list or other sparse representation of network data into an adjacency matrix format. For example, the **igraph** package includes the `as.matrix.igraph` function, and the **sna** package includes the `as.sociomatrix.sna` function.

> The method provides estimation of a one mode (i.e. unipartite) symmetric (i.e. undirected) longitudinal network data. In order to analyze a bipartite network data set, users may apply one-mode projection of each cross-sectional network to obtain a symmetric matrix for a node set which belongs to one of the two modes.
