### 2021-11-30 jongheepark/NetworkChange#8 @ysohn
The following is included in the paper.

> The algorithm's computational cost is heavily influenced by the size of the input network data set as well as the number of dimensions and breaks to be estimated. In most cases, two-dimensional estimation will provide a reasonable description of the latent factors. When using a personal computer and the current version of the package, two-dimensional estimation for a single break results the following execution time for the corresponding longitudinal network size.

> A graph of execution time for longitudinal networks with different size.

> Because the number of breaks increases the number of states and state-specific latent factors to be estimated, increasing the number of states has a greater impact on `NetworkChange` function’s execution time than the other factors. Please note that `BreakDiagnostic`, which runs `NetworkChange` over different numbers of breaks, spends time according to the length of time spent by `NetworkChange` given the aforementioned conditions.
