### 2021-11-30 jongheepark/NetworkChange#5 @ysohn
If the user does not set constant = TRUE, which draws intercept samples, the B0 and b0 parameters are not used in MCMC sampling. The parameters B0 and b0 are included in the code for internal testing and future use.
