### 2021-11-30 jongheepark/NetworkChange#7 @ysohn
We have expanded the paper by reflecting the reviewers and the editor's comments.
