### 2021-12-02 jongheepark/NetworkChange#6 @ysohn
We appreciate the reviewer for highlighting the need of convergence diagnostics. As shown in your analysis, the proposed model contains a large number of sampled parameters, and we need to concurrently monitor their convergence. When monitoring, it is important to notice that several conventional transformation procedures (e.g. Gram–Schmidt orthogonalization for U) are applied following the completion of each MCMC iteration.

For the primary parameters used to represent the relational features of nodes on each state, U, the default option for normalization is Gram–Schmidt orthogonalization which can result in a combinatorially huge number of U column sets. Thus, for each state, the N x R matrix U is only identifiable up to UU', where U' denotes the transpose of U matrix.

Simulation example:
https://htmlpreview.github.io/?https://github.com/ysohn/reviews/blob/main/JOSS_NC/diag_report.html
When comparing across 8 chains in the case shown, all parameters examined (UU' instead of U) converge well, with rhat values less than 1.05, the threshold of convergence proposed by Vehtari et al. In this case, we must keep track of N x (N-1)/2 x (m+1) off-diagonal elements for the symmetric matrix UU' of all m+1 states and R x T elements for V.

While this simulated example with simple ground truth parameters does not demonstrate a significant multi-modal structure, it is not surprising that real-world data generating processes are more likely to be the result of multiple heterogeneous factors that produce a multi-modal posterior landscape in both temporal and state-specific group structural dimensions. Even static empirical networks are known to have multi-modal structures that are difficult to understand by aggregating stochastic estimates of their partitions (Peel, Larremore, Clauset, 2017).

When we apply our method to the major alliance example, over repeated trials, 60 to 70% of the approximately 200 monitored parameters converge quickly, exhibiting less than 1.05 rhat values. However, a subset of the parameters do not pass the threshold after a moderately long number of iterations for the eight chains studied. This may be due to the aforementioned nature of heterogeneous factors underlying the generation of the observed networks, so that it takes more time to travel distant modes in the high dimensional parameter space. This, conversely, given the straightforward result of the simulation example, implies that the result for the major alliance network may be due to data generating factors that do not align with each other.

Given that we use highly noninformative priors, multiple chain swapping approaches for recovering a multimodal target density, such as parallel tempering, may be helpful for a more efficient recovery of the multi-modal structure in the high dimensional parameter space. We leave this possibility for future research. As a guidance, we mentioned in the paper that the users may run multiple chains to explore modes that are distant to each other unless they have enough computing resources for a single MCMC chain to recover a multi-modal posterior landscape.
