### 2021-12-06 jongheepark/NetworkChange#3 @ysohn
We understand the reviewer's suggestion that it would be good to have an example to show how well the model recovers planted parameters of a data set generated with a known data generating process. However, we should say that it is difficult for the case of our model as several transformation techniques, such as the null function subtraction, are applied to the original input network data set, and also due to the fact that there is not a consensus in network science about what should be a general network generating model with tunable parameters to represent many forms of networks (our model's noninformative priors can never serve this purpose).

Instead of explicitly recovering ground truth parameters, NetworkChange recovers key characteristics of an input data set in the latent space domain. For instance, each regime-specific U and V reveals the interested group (or mesoscopic) structure associated with the corresponding regime's network topology as shown by the two dimensional plots of node embeddings and their clustering. As a result, UVU' represents the characteristic features (e.g. block structures) of the networks belonging to the corresponding regime, rather than re-generating the observed adjacency matrices or other planted parameters.

If we instead say that the fundamental parameters of interest are the timings of change or the forms of group structure, we believe that the current version recovers those in a reasonable extent as shown in the simulation example. However, the data generating equations for the input networks are, in general, undefined, so that the recovery performance would be highly subject to specific network generating models and changes employed among numerous possibilities. Due to these reasons, we should say that it is difficult to meet the reviewer's request.
