### 2021-12-19 #2708 @martinmodrak
Hi, I did skim your responses and I think that (as I said before) the package being transparent about convergence (or lack of it) is what's important. I'll try to read and respond in depth, but I won't be able to have a response ready before Christmas break.
