### 2022-02-07 #2708 @martinmodrak
Sorry for the delay on my side. I finally got to review the changes made.

I still don't think a change to the underlying MCMC method is necessary/recommended for JOSS acceptance, but I think the current version of the package sets the users up for failure in several ways, neither of which was fixed by the update (which seems to have changed little beyond the JOSS paper itself):

1) The potential limitations of the MCMC method and limitations on identifiability and convergence of the model are mentioned only in the JOSS paper, but not in the documentation for `NetworkChange`, neither in the example vignette.
2) It is hard for users to verify whether the MCMC has converged and whether the posterior is multimodal, making it impossible to check that the inferences they'll make from their datasets are valid. In particular:

    a) There should be a straightforward way to assess convergence of the relevant parameters. Rhat/ESS is an option I would recommend, but there definitely are others which the authors may choose to use. Having a simple way to show traceplots would also be nice (but not necessary).

    b) the `NetworkChange` method should by default issue a warning if any of the parameters has bad diagnostic values

    c)  Checking the diagnostics should be shown in the examples in the vignette.

    d) the `NetworkChange` method should accept a parameter to specify the number of chains to run, with default of at least 4. (suggestion: use the `future` package to run the individual chains, which will let you very easily allow users to run the chains in parallel).

3) The default number of burnin and mcmc samples is very low. This means that non-convergence due to short warmup is more likely and only low precision for any variable (including the MCMC diagnostics) is achieved. Note that 100 samples give me quite low precision, even if the samples were completely uncorrelated (which they aren't). E.g.  if I take 100 i.i.d. samples from a normal distribution with mean 0 and sd 1, the observed mean will fluctuate between -0.2 and 0.2 , observed sd will be 0.86 - 1.13 and observed 0.975th quantile would be 1.44 - 2.38 (all are 95% intervals).

I think those are major shortcomings that need to be addressed. The ordering above is roughly the order of importance I see in those issues. I believe extending the documentation is easy and important and is IMHO a blocking issue for acceptance. Adding a way to access the diagnostics is also relatively easy and is IMHO a high priority, although not a 100% must. I feel less strongly about adding multiple chain support and changing the default sample numbers, so if there is a good reason not to do those, I am open to be convinced so.

Additionally, I think there is a misconception in the reply to https://github.com/jongheepark/NetworkChange/issues/6 as well as in the paper. I discuss that in my comment on the issue and expect to either be proven wrong or the wording in the paper to change accordingly.
