+++ vignettes/NetworkChange.bib
@@ -1,12 +1,20 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for 박종희 at 2020-07-12 03:01:37 +0900 
+%% Created for 박종희 at 2020-11-05 23:29:56 +0900 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@book{Gibler2009,
+	Author = {Gibler, Douglas},
+	Date-Added = {2020-11-05 23:27:24 +0900},
+	Date-Modified = {2020-11-05 23:27:24 +0900},
+	Publisher = {CQ Press},
+	Title = {{International Military Alliances, 1648-2008}},
+	Year = {2008}}
 
 @article{Park2020,
 	Author = {Jong Hee Park and Yunkyu Sohn},
