+++ NAMESPACE
@@ -8,6 +8,7 @@ export(NetworkChange)
 export(NetworkChangeRobust)
 export(NetworkStatic)
 export(ULUstateSample)
+export(ULUstateSample.mpfr)
 export(WaicCompare)
 export(drawPostAnalysis)
 export(drawRegimeRaw)

[…]
