+++ DESCRIPTION
@@ -1,21 +1,21 @@
 Package: NetworkChange
 Title: Bayesian Package for Network Changepoint Analysis
 Version: 0.8
-Date: 2022-02-24 14:17:31 UTC
+Date: 2022-03-03 14:17:31 UTC
 Author: Jong Hee Park [aut,cre], Yunkyu Sohn [aut]
 Maintainer: Jong Hee Park <jongheepark@snu.ac.kr>
 Depends: R (>= 2.10.0), MCMCpack, ggplot2
 Imports: grid, Rmpfr, abind, mvtnorm, tidyr, sna, igraph, qgraph, network,
 	 stats, MASS, methods, RColorBrewer, reshape, ggrepel, gridExtra, rlang, 		 GGally, ggvis
-Description: Network changepoint analysis for undirected network data. The package implements a hidden Markov network change point model (Park and Sohn 2020 <doi:10.1214/19-BA1147>). Functions for break number detection using the approximate marginal likelihood and WAIC are also provided.
+Description: Network changepoint analysis for undirected network data. The package implements a hidden Markov network change point model (Park and Sohn (2020)). Functions for break number detection using the approximate marginal likelihood and WAIC are also provided.
 License: GPL-3
 NeedsCompilation: no
-Packaged: 2022-02-24 14:17:31 UTC
+Packaged: 2022-03-03 14:17:31 UTC
 RoxygenNote: 7.1.1
 Roxygen: list(markdown = TRUE)
 Repository: CRAN
 Encoding: UTF-8
-Date/Publication: 2022-02-24 14:17:31 UTC
+Date/Publication: 2022-03-03 14:17:31 UTC
 Suggests: 
     knitr,
     rmarkdown
